+++ driver/web/apis_admin.go
@@ -220,7 +220,7 @@ func (h AdminApisHandler) getAccounts(l *logs.Log, r *http.Request, claims *toke
 	if err != nil {
 		return l.HttpResponseErrorAction(logutils.ActionGet, model.TypeAccount, nil, err, http.StatusInternalServerError, true)
 	}
-	response := getAccountListToDef(getAccounts)
+	response := getAccountsListToDef(getAccounts)
 
 	data, err := json.Marshal(response)
 	if err != nil {
@@ -242,9 +242,9 @@ func (h AdminApisHandler) getAccount(l *logs.Log, r *http.Request, claims *token
 		//permissions
 		permissions := applicationPermissionsToDef(account.Permissions)
 		//roles
-		roles := appOrgRolesToDef(account.Roles)
+		roles := appOrgRolesToDef(nil)
 		//groups
-		groups := appOrgGroupsToDef(account.Groups)
+		groups := appOrgGroupsToDef(nil)
 		//account auth types
 		authTypes := accountAuthTypesToDef(account.AuthTypes)
 		accountData = &Def.ResSharedAccount{Id: account.ID, Permissions: &permissions, Roles: &roles, Groups: &groups, AuthTypes: &authTypes, Profile: profile}
@@ -256,6 +256,31 @@ func (h AdminApisHandler) getAccount(l *logs.Log, r *http.Request, claims *token
 	}
 
 	return l.HttpResponseSuccessJSON(data)
+
+}
+func (h AdminApisHandler) accountExists(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HttpResponse {
+	data, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionRead, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, false)
+	}
+
+	var requestData Def.ReqAccountExistsRequest
+	err = json.Unmarshal(data, &requestData)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionUnmarshal, logutils.TypeRequest, nil, err, http.StatusBadRequest, true)
+	}
+
+	accountExists, err := h.coreAPIs.Auth.AccountExists(string(requestData.AuthType), requestData.UserIdentifier, requestData.ApiKey, requestData.AppTypeIdentifier, requestData.OrgId, l)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionGet, logutils.MessageDataType("account exists"), nil, err, http.StatusInternalServerError, false)
+	}
+
+	respData, err := json.Marshal(accountExists)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionMarshal, logutils.TypeResponse, nil, err, http.StatusInternalServerError, false)
+	}
+
+	return l.HttpResponseSuccessJSON(respData)
 }
 
 //NewAdminApisHandler creates new admin rest Handler instance