+++ driver/web/conversions_user.go
@@ -172,37 +172,6 @@ func profileFromDefNullable(item *Def.ReqSharedProfileNullable) model.Profile {
 		State: state, Country: country}
 }
 
-//Accounts
-func getAccountsToDef(item *model.Account) *Def.ResSharedAccount {
-	if item == nil {
-		return nil
-	}
-	profile := profileToDef(&item.Profile)
-	//permissions
-	permissions := applicationPermissionsToDef(item.Permissions)
-	//roles
-	roles := appOrgRolesToDef(item.Roles)
-	//groups
-	groups := appOrgGroupsToDef(item.Groups)
-	//account auth types
-	authTypes := accountAuthTypesToDef(item.AuthTypes)
-	return &Def.ResSharedAccount{Id: item.ID, Permissions: &permissions, Roles: &roles, Groups: &groups, AuthTypes: &authTypes, Profile: profile}
-
-}
-
-func getAccountsListToDef(item []model.Account) []Def.ResSharedAccount {
-	out := make([]Def.ResSharedAccount, len(item))
-	for i, item := range item {
-		defItem := getAccountsToDef(&item)
-		if defItem != nil {
-			out[i] = *defItem
-		} else {
-			out[i] = Def.ResLoginAccount{}
-		}
-	}
-	return out
-}
-
 //Device
 func deviceFromDef(item *Def.DeviceFields) *model.Device {
 	if item == nil {