+++ CHANGELOG.md
@@ -92,6 +92,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Expose get accounts admin API [#283] (https://github.com/rokwire/core-building-block/issues/283)
 - Expose get account admin API [#270] (https://github.com/rokwire/core-building-block/issues/270)
+- Expose does account exist admin API [#271](https://github.com/rokwire/core-building-block/issues/271)
 - Extract IP address from request [#246](https://github.com/rokwire/core-building-block/issues/246)
 - Populate profile data automatically from ROKWIRE 2.0 Profile BB [#185](https://github.com/rokwire/core-building-block/issues/185)
 - Add phone authentication support [#24](https://github.com/rokwire/core-building-block/issues/24)